Set up Antidote for zsh
Realized I was only using a couple parts of oh-my-zsh, so I decided to
rip it out and just source those files and use Antidote to manage my zsh
plugins. https://getantidote.github.io/

As a bonus, this means I no longer need to use Git submodules in this
repo to load oh-my-zsh and zsh-autosuggestions.